[Mobile] clean codes
