+++ rideshare/lib/constants/json.dart
@@ -1,0 +1,27 @@
+final List<Map<String, String>> historyList = [
+    {
+      'SourceIcon': '4 Kilo Abrehot Library',
+      'DestinationIcon': 'Megenagna Metebaber BLDG',
+      'DateTime': '5th June 2023 12:28 AM',
+      'Price': 'Br 60',
+    },
+    {
+      'SourceIcon': '4 Kilo Abrehot Library',
+      'DestinationIcon': 'Megenagna Metebaber BLDG',
+      'DateTime': '5th June 2023 12:28 AM',
+      'Price': 'Br 60',
+    },
+    {
+      'SourceIcon': '4 Kilo Abrehot Library',
+      'DestinationIcon': 'Megenagna Metebaber BLDG',
+      'DateTime': '5th June 2023 12:28 AM',
+      'Price': 'Br 60',
+    },
+    {
+      'SourceIcon': '4 Kilo Abrehot Library',
+      'DestinationIcon': 'Megenagna Metebaber BLDG',
+      'DateTime': '5th June 2023 12:28 AM',
+      'Price': 'Br 60',
+    },
+    // Add more history items as needed
+  ];
